ftp: fix labels in /var/lock/subsys/
In the pattern "/var/lock/subsys/*.ftpd", the star is applied to the
slash instead to the dot.  This means that the pattern matches these
files:

* "Xftpd" in "/var/lock/subsys/" (where X is whatever character)
* "subsysXftpd" in "/var/lock/" (where X is whatever character)

"/var/lock/subsys/vsftpd", which has been used by vsftpd, is therefore
not matched by the pattern.

As "*." looks like a typo, this patch replaces it with ".*".