pulseaudio: update server and client permissions
Update the pulseaudio module, so that:
- the pulseaudio server can send the signull signal to its clients;
- the pulseaudio server can manage not only pulseaudio_tmpfs files
  but all pulseaudio_tmpfs_content;
- pulseaudio clients can also write and not just read pulseaudio
  content files in the temporary filesystems (tmpfs);
- pulseaudio clients can use file descriptors for pulseaudio.

This patch has been triggered by recent testing of the games module
and it makes the following patch obsolete (not applied yet):

[PATCH 2/2] pulseaudio: improve the support for Gnome games
http://oss.tresys.com/pipermail/refpolicy/2016-December/008651.html

It also makes obsolete interface calls to the following interfaces
in addition to pulseaudio_run(): pulseaudio_rw_tmpfs_files() and
pulseaudio_use_fds(). These are currently used in the mozilla
module (removed in part 2/2 of this patchset).

Signed-off-by: Guido Trentalancia <guido@trentalancia.net>